[export_ps] -Fix: packed color really needs to be at least 32 bit unsigned int for portability
git-svn-id: svn://repo.hu/pcb-rnd/trunk@27028 1d61de99-560f-4dba-8957-1e505d42358b

+++ src_plugins/export_ps/eps.c
@@ -31,7 +31,7 @@ typedef struct hid_gc_s {
 	pcb_core_gc_t core_gc;
 	pcb_cap_style_t cap;
 	pcb_coord_t width;
-	int color;
+	unsigned long color;
 	int erase;
 } hid_gc_s;
